Separated out logic from slide
